collections/ansible_collections/lv3/platform/roles/piper_runtime/files/requirements.txt
piper-tts[http]==1.4.1
